backend/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "mistralai-narutoverse-chatbot"
version = "0.1.0"
description = "NarutoVerse Chatbot powered by MistralAI (SWE Internship Project)"
authors = ["Mudassar Zahid <mudassar.ahmad.zahid@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.0"
python-dotenv = ">=1.0.1,<1.1.0"
fastapi = ">=0.115.2,<0.116.0"
requests = ">=2.32.2,<2.33.0"
httpx = ">=0.27.2,<0.28.0"
tqdm = ">=4.66.5,<4.67.0"
langchain-mistralai = "^0.2.0"
sqlmodel = ">=0.0.22,<0.1.0"
beautifulsoup4 = ">=4.12.3,<4.13.0"
sqlalchemy = ">=2.0.35,<2.1.0"
pydantic = ">=2.9.2,<2.10.0"
langchain = ">=0.3.3,<0.4.0"
langgraph = ">=0.2.38,<0.3.0"
typing-extensions = ">=4.12.2,<4.13.0"
starlette = ">=0.39.2,<0.40.0"
ujson = "^5.10.0"
langchain-chroma = "^0.1.4"

[tool.poetry.group.dev.dependencies]
types-requests = "^2.32.0.20241016"
types-beautifulsoup4 = "^4.12.0.20241020"
types-tqdm = "^4.66.0.20240417"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
pytest-html = "^4.1.1"
black = "^24.10.0"
flake8 = "^7.1.1"
mypy = "^1.12.1"
pydocstyle = "^6.3.0"
isort = "^5.13.2"

# pytest
[tool.coverage.run]
omit = ["tests/*", "__init__.py"]

# Linting
[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pydocstyle]
convention = "google"
add-ignore = "D100,D104"

[tool.mypy]
python_version = "3.10"
warn_redundant_casts = "True"
warn_unused_ignores = "True"
warn_return_any = "True"
warn_unused_configs = "True"
warn_unreachable = "True"